+++ packages/react/src/sds/surveys/SurveyFormBuilder/QuestionTypes/DropdownSingleQuestion/types.ts
@@ -8,6 +8,7 @@ export type DropdownSingleQuestionOnChangeParams =
     datasetKey?: string
     showSearchBox?: boolean
     searchBoxPlaceholder?: string
+    allowCreate?: boolean
   }
 
 export type DropdownSingleQuestionProps =
@@ -17,4 +18,5 @@ export type DropdownSingleQuestionProps =
     value?: string | null
     showSearchBox?: boolean
     searchBoxPlaceholder?: string
+    allowCreate?: boolean
   }